+++ db/migrate/20160422145449_rename_question_number_to_position.rb
@@ -1,0 +1,5 @@
+class RenameQuestionNumberToPosition < ActiveRecord::Migration
+  def change
+    rename_column :survey_question_orders, :question_number, :position
+  end
+end